Перемещена функция 'shuffle' и константы
